refactor(rules_toolchain): simplify toolchain providers
This patch removes an extra layer of indirection in toolchains.

The `toolchain/local/binary` rule sets up the `TemplateVariableInfo`,
`ToolchainInfo` and `DefaultInfo`.

The newly renamed `resolved` rule just forwards on the created
providers after the toolchain resolution has happened.

This removed the need to create an extra `info` rule
to describe the toolchain