fix(rules_toolchain): add `no-remote` to local binary symlinks
It is almost impossible to correctly transfer all the local files necessary
for a local binary to the remote service.

For safety, constraint all local toolchains to run on the local client until we
have more time to validate that local toolchains work remotely.

This is certainly possible: C/C++ toolchain define all the local files needed
for remote execution.

We will need to expose the `data` attribute through so that arbitary files can
be globbed and sent along with the binary.