refactor(rules_toolchain): shift to `use_repo_rule`/`visibility`
`use_repo_rule` is available in `6.4.0` and above.

It _hugely_ simplifies what we want to do.

That should be the new minimum version for our rulesets.

As of writing, `6.4.0` is not released. The release candidates do not have
the `use_repo_rule` added, yet. Most rules use `last_rc` Bazel version.

`visibility` is available in 6.0+.