feat: add a `resolved` target to local binary repositories
This removes the need for a custom `resolved` rule for every toolchain.

Upgrade path is to add the `toolchain_type` to the local repository
rule in `MODULE.bazel` and change the `resolved` target to an alias.

A `toolchain_resolved` repository rule is added as well for toolchains
that do not use local binaries.

This patch hugely simplifies the creation of toolchains downstream.