feat: add launcher
This is a bit of a hack for Windows.

We want to be able to do:

    bazelisk run -- @echo//:entrypoint

However, because there is no extension on `entrypoint` which is
actually a hardlink/symlink to `entrypoint.bat` Bazel puts it through
`CreateProcessW` and attempts to start it as a portable executable
(PE/COFF).

The launcher is a bit of C# code to create an `exe` that uses argument
zero to find the corresponding `.bat` script and launches that instead.

The resulting executable can be hardlinked/symlinked from anywhere that
has a batch script to provide a executable launcher for the script.

We can then have nice labels on both POSIX/Windows.