fix(symlink/target): add `files_to_run` property on `ToolchainInfo` provider
BREAKING CHANGE: `toolchain_symlink_path` no longer provides `ToolchainInfo`/`TemplateVariableInfo`.

It is required to use `toolchain_symlink_target` to the `toolchain_symlink_path`.

This is due to `toolchain_symlink_target` needing to export the `DefaultInfo.file_to_run` from the
symlinked target. That is not possible with a filesystem path.

This should not affect most users as `toolchain_symlink_path` is mainly used in `toolchain_local_which`.

If `toolchain_symlink_path` is being used, then to gain the previous functionality, create a
`toolchain_symlink_target` to the `toolchain_symlink_path` label.